+++ app/helpers/request_helper.py
@@ -24,7 +24,10 @@ def remote_call(url, json=None):
 
 
 def response_ok(response):
-    if response.status_code == 200:
+    if response is None:
+        logger.error("No response from service")
+        return False
+    elif response.status_code == 200:
         logger.info("Returned from service", request_url=response.url, status_code=response.status_code)
         return True
     else: