+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Unreleased
+  - Fix handling of None responses in remote call
   - Add SFTP support
 
 ### 1.0.0 2017-02-2016